+++ lib/bootstrap_form/helpers/bootstrap.rb
@@ -51,18 +51,21 @@ module BootstrapForm
       end
 
       def prepend_and_append_input(options, &block)
-        options = options.extract!(:prepend, :append, :prepend_button, :append_button)
+        options = options.extract!(:prepend, :append)
         input = capture(&block)
 
-        input = content_tag(:span, options[:prepend], class: 'input-group-addon') + input if options[:prepend]
-        input << content_tag(:span, options[:append], class: 'input-group-addon') if options[:append]
-
-        input = content_tag(:span, options[:prepend_button], class: "input-group-btn") + input if options[:prepend_button]
-        input << content_tag(:span, options[:append_button], class: "input-group-btn") if options[:append_button]
-
+        input = content_tag(:span, options[:prepend], class: input_group_class(options[:prepend])) + input if options[:prepend]
+        input << content_tag(:span, options[:append], class: input_group_class(options[:append])) if options[:append]
         input = content_tag(:div, input, class: "input-group") unless options.empty?
-
         input
+      end
+
+      def input_group_class(add_on_content)
+        if add_on_content.match /btn/
+          'input-group-btn'
+        else
+          'input-group-addon'
+        end
       end
 
       def static_class